requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.4
Flask-MySQLdb==1.0.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
